Some probabilities become so low, that they turn out to be zero. These are:
P(u) ~ product over user tags can be very low
P(u|i) ~ same reason

Solutions:
  - For the relevant items estimate, I had to pre-compute the log and
perform a
summation
  - For the entropy, I ignored P(u). I can't see what else can be
done.

Also changed in this revision:
  - I optimized the code a bit with the cost of extra memory. This was
as
consequence of another issue which is: When in the baseline, we filter
out some
of the user items, these annotation filtered should not be used in
probability
estimates. They were.